Updated openstack/openstack Project: openstack/neutron  5447f7e427af68555366515d5153760ae9fc7e9c
Do not use auto_schedule_routers to add router to agent

auto_schedule_routers makes a number of DB calls that are
unnecessary as they are already made during the validation
phase of add_router_to_l3_agent. Once the validation is
done, the only business left is to create the binding: this
is what this patch does.

Partial-bug: #1356121

Change-Id: Ia9be998c2b94416bc46ef78415099f0099b08c2a