Updated openstack/openstack Project: openstack/ceilometer  f49c648cd1157541d0e7720d01c73bc78d3325ba
Adding comparison operators in query for event traits

In EventFilter we have an optional parameter 'op' in traits_filter part. But up
to date usage of this param has been implemented only in Hbase and
Mongodb backends. However in API we can use only 'op'='eq'.

This change improves usage of parameter 'op' in query for events in api, and
adds possibility to use comparison in query with SQL backend.

Some tests are also added to cover using of this new feature.

Closes-bug: #1354009
Change-Id: I4824800eac7270bcfa10882a735e48d80bf32712