Updated openstack/openstack Project: openstack/cinder  e5e61bf7ce11aa23b48fa22574c8d4e1c1ea9a7d
Fixing 3PAR excessive FC port usage

Updating the 3PAR FC driver so that it can detect if there is
only a single FC path available.  When a single FC path is
detected only a single VLUN will be created instead of one for
every available NSP on the host.  This will prevent a host
from using extra FC ports that are not needed.  If multiple
FC paths are available all the ports will still be used.

Closes-Bug: 1360001
Change-Id: I4b2bdf12289dd255dd959675991fa9e2e0fcbd1e