Updated openstack/openstack Project: openstack/python-novaclient  3fa05957daa44de7d16e0f1930b36dd6791dc1cd
Add support for security-group-default-rules

Add API & CLI support for security-group-default-rules. Security group
default rules allows users to create, delete, list, and get rules that
will be pre-populated into the Security Group "default" that is
populated in all projects on creation. This feature was intoduced in
Grizzly.

This commit also includes a small typo fix in the security_group_rules fake.

Change-Id: I6722b659934bde34bd206103f92dbd2cc42b1537
Closes-Bug: #1326941