Updated openstack/openstack Project: openstack/keystone  c8ba37089f1dcf70c3eb60ffa50fb646995f6300
Mark methods on token_api deprecated

Mark the methods on token_api as deprecated. The versionutils.deprecated
decorator will only work with functions and classes (meaning it cannot
be applied to instance methods post instantiation). A log line is used
that should only trigger once per object on token_api instead of the
versionutils decorator.

Change-Id: I97c96c239262697b0df37ef895c4bafa262892e0
bp: non-persistent-tokens
bp: deprecated-as-of-juno