Updated openstack/openstack Project: openstack/heat  7b40b2e96756012c28579a9b20ffd7a890ac79a8
Push metadata changes to the TempURL

This changes EngineService._push_metadata_software_deployments
to push deployments metadata changes to any configured
TempURL as stored in the server resource data.

The server resource metadata is treated as the canonical
source of up-to-date metadata, so deployments data is always
merged with existing resource metadata before persisting to
the database.

Implements-Blueprint: swift-deployment-transport
Change-Id: I96698ac5796c3c5c4381750d9a6f2d6a790fe25b