Updated openstack/openstack Project: openstack/cinder  a02cc07bd823e9795b1563ce058772795772bbd9
Add support for backup encryption metadata

This modification allows the encryption key UUID field, which
has been added to the volume table, to remain valid when encrypted
volumes are backed up and then restored, which enables the restored
volume to be accessible and encrypted.  This is related to patch
https://review.openstack.org/#/c/39573/, except that it uses the
backup metadata support recently added in patch
https://review.openstack.org/#/c/51900/ rather than modifying the
backup api.

Change-Id: Ib91f8275271e6bd4f2d9e17499d16ec13bca2b84
Implements: blueprint backup-support-for-encrypted-volumes
DocImpact