Updated openstack/openstack Project: openstack/nova  19e868bf3e52a6d80dc8c7ef8f93b0ff391cba29
libvirt: reworks configdrive creation

This refactor the creation of the configdrive by using
the same code schema as local/ephemeral/swap/disk.

Now this is the imagebackend that creates the configdrive like any other
disk attached to VM. This ensures that the configdrive file is created
where the imagebackend expect it.

This also removes the assumption that the configdrive was always created
at the right place, that was not true for rbd and lvm.

Related-Bug: #1361840

Change-Id: Ia03a83d593794c8b2a644360790cb15de9dea6e9