Updated openstack/openstack Project: openstack/nova  b26a281ebe65f9bc2bfa1488856b6869643e98a6
libvirt: Add method for getting host NUMA topology

We can get this information from the libvirt capabilities XML now, so we
add a method that returns it in the form of
virt.hardware.VirtNUMATopology objects. We will exclude any CPUs not
found in the CONF.vcpu_pin_set.

Change-Id: Ibfc60bc71e49579bc5ab4996b752ed7a49830b6a
Blueprint: virt-driver-numa-placement