Updated openstack/openstack Project: openstack/taskflow  70d9199ea374b558a9918089de94af0b6c76811b
Use checked_commit() around consume() and abandon()

To ensure we reliably handle when a transaction fails
we should use the checked_commit() helper function instead
of the currently not fully exception handling kazoo
transaction commit function.

This should be addressed in the future with:

- https://github.com/python-zk/kazoo/pull/224
- https://github.com/python-zk/kazoo/pull/225

Those have not merged yet (or been released) so we need
to use a similar function in the meantime.

Change-Id: Icf83b7d4955c11227e733287170a7bd3ab372bd2