Updated openstack/openstack Project: openstack-dev/devstack  9a59e81eb0a8b5687f7f982006a0c401993f2d04
Add graphviz package to general list for building docs

A few projects are now gating on docs not having warnings/errors in them
during the docs build with sphinx, so developers are going to want to be
able to run 'tox -e docs' locally and fix issues.

The graphviz package is used by sphinx when building docs but you have
to get it from the distro, so this change adds the package to the
various distro lists.

Closes-Bug: #1359749

Change-Id: I7b7dee5e5d0dce6d5020b3e711f500118be163f8