Updated openstack/openstack Project: openstack/heat  9abf5f9514e86da8564fc18d4680bded384e32eb
Patch 044 migration to work w/ MySQL 5.6

With MySQL 5.6 setup to use UTF-8 for everything, the 044 migration
will fail. This is because a VARCHAR(256) cannot be indexed in UTF-8
databases. In versions of MySQL before 5.6, this produced a warning and
a truncated, less-useful index. In 5.6 it was changed to an error.

We will reduce these columns further to 64 bytes in the model in a
subsequent fix. For now, we will just create the table with 64 character
columns on new systems, which will allow this migration to work properly
on mysql-5.6. There is no danger of overflow, as 64 chars is what Keystone
uses for them anyway.

Change-Id: I73932f69ff662ef5bb030b6516bd3d32507de658
Related-Bug: #1316646