Updated openstack/openstack Project: openstack-infra/devstack-gate  fc6c079595cb8933452d921ace549233edaa38d6
Add features.yaml entry for 'no services'

Previously service selection was only additive on top of the default
set of services for a given branch.  In order to support more
efficient functional job definitions, this change adds the option
of setting DEVSTACK_GATE_NO_SERVICES to ensure that the default
set is minimal (mysql, rabbit, dstat at the time of this patch).

Change-Id: Ie20f611b5c5debe96df1001df5d19eb24b93c519