Updated openstack/openstack Project: openstack/keystone  1ac1de809c267a81fed3dcf6a058afa3348a8aeb
Add index for actor_id in assignments table.

Searching the assignments table by actor_id is a common occurance
and in a large OpenStack installation, this can be a significant
performance degradation.  This table is read much more often than
it is written, so this should be a good trade off.

Change-Id: Ia1a4392ca2118690a10130b10b2f69cd05672a12
Closes-bug: 1362557