Updated openstack/openstack Project: openstack/neutron  b943e31b26cd23dea48899030c279b0cc5ab4007
NSX: log request body to NSX as debug

Previously we never logged the request body that we sent to NSX. This makes
things hard to debug when issues arise as we don't actually log the body of
the request that we made. This patch adds the body to our issue request log
statement.

Change-Id: I7d98e3d844ebe565fb5919aac3f9afd5b4040e69
Closes-bug: 1360394