Updated openstack/openstack Project: openstack/keystone  647193db685ffff4cff8a41f1e4dd93d53232538
Fix token flush fails with recursion depth exception

keystone-manage token_flush was failing with a recursion depth
exception error. This is because when a property was referenced in
token.persistence.Manager it would go into an infinite recursion
trying to resolve it. token.persistence.PersistenceManager should
have been used instead.

Note that I added a test for keystone.cli TokenFlush, but for some
reason it doesn't actually expose the problem. I'll leave it there
since it increases coverage.

Change-Id: If5af4fe0811a59157c2f68d06c264815409689e1
Closes-Bug: #1363224