Updated openstack/openstack Project: openstack/nova  1f677b0bf6f6312f89126e5d57b7fcbbed812bd2
Adds Hyper-V serial console log

Introduces the VM serial output log feature to the Hyper-V Nova
 Compute driver.

The VM serial port COM0 is redirected to a named pipe so that a
worker can take the VM output and write it to a file.

The VM serial output is then read from this file whenever the
console log command is issued.

Implements: blueprint hyper-v-console-log

Co-Authored-By: Adelina Tuvenie <atuvenie@cloudbasesolutions.com>

Change-Id: I7d3132c0f934ac3e8d28315f4da766466612a6c0