Updated openstack/openstack Project: openstack/horizon  61191da6e15b37c8772af070569356fce3dcd43c
Disable Floating IP features if Neutron router is disabled

If the config option 'enable_router' is set to False,
Floating IP features are disabled when Neutron is enabled.
It does not affect when Neutron is disabled.

It also adds unit tests for api.network.servers_update_addresses
which is affected by this change.

Completes blueprint hide-router-panel-by-config
Closes-Bug: #1292022
Change-Id: Ib63c6a0e7bb5661d4a60d10a1722fdad978b50bb