Updated openstack/openstack Project: openstack/horizon  9416662e42230b1e81a88f002d760fa8b126d72f
Upload volume to image service

Expose the functionality of the 'cinder upload-to-image' command
in the UI. It allows user to upload of a volume whose status is
in-use or available to the Glance image service.

When the volume is in-use, user still wants to uploads that volume
to image, user needs to set force to be True. The force checkbox
only shows when the volume is in-use. Whether it can
succesfully upload the volume to image depends on if the storage
array that volume is created in supports attaching multiple instances
to the volume. Not all arrays supports that. There is no api that
can detect if array supports that before cinder actutally uploads
the volume in-use to image.

The container format is 'bare'. The list of disk formats are
based on the format supportability by both glance and qemu-img.
cinder uses qemu-img for converting the volume to a image.

Implements: blueprint cinder-volume-upload-image

Change-Id: Ie5fc26c260e5f4ef2700c40c8cea6150fdbd522c