Updated openstack/openstack Project: openstack/nova  6def23d4b0f77d74cd06b3db750fd57a98345223
Check min_ram and min_disk when boot from volume

Currently when boot from volume and the volume is created
from a image, the original image's min_ram, min_disk attributes
are ignored, this is not good.

We should extract these properties from the volume and check
it when boot from volume.

Change-Id: I861a78b5c7efa71e4bf7206d388b8d0d8048c78e
Closes-Bug: #1360022