Updated openstack/openstack Project: openstack/neutron  f1be6108def6a124bc3e3a03dceb41918ce17f11
Remove SELECT FOR UPDATE use in update_firewall

SELECT FOR UPDATE expression, which is triggered with the use of the
SQLAlchemy Query object's with_lockmode('update') method, is
detrimental to performance and scalability of the database
performance code in Neutron due to the lock contention it produces.

SELECT FOR UPDATE can be entirely avoided in update_firewall method
with the use of single-shot UPDATE expressions.

Change-Id: I333f726498192295116ca96ff5ca9012c36a6fd1