Updated openstack/openstack Project: openstack/horizon  8e11b803f650ada7abac88ab7182fe5304c1279e
Display Network panels based on neutron ext-list

Neutron features like LBaaS, FWaaS or VPNaaS are provided as
extensions in Neutron. This commit changes these features are enabled
only when they are included in the extension list from Neutron.

It also change the default settings of enable_lb/firewall/vpn to True.
The default of these settings were set to False to avoid confusion to
users because LB/FW/VPNaaS are optional features in Neutron. By this
commit corresponding features in Horizon are enabled dynamically, so
it is reasonable to change the default to True.

This commit introduces Horizon version of override_settings
"update_settings" decorator which allows override an item in dict.
Django original override_settings replaces a dict completely,
however OpenStack dashboard setting has many dictionary configuration
and there are test case where we want to override only one item
in a dictionary and keep other items in the dictionary.
This version of override_settings allows this.

DocImpact

Implements blueprint neutron-feature-control-with-ext-list

Change-Id: I19424d448c800524de0ee8ac769974272dbb84a0