Updated openstack/openstack Project: openstack-dev/devstack  0a8c555079fefdbff73852df38fa513a66d0bea4
Faster nova fixed key generation

Using bc 64 times in loop is too verbose and slow,
replacing the echo/bc loop with hexdump and urandom.

The hexdump approach is 75 times faster and
does not floods the debug logs.

Using the common function for generating,
this kind of string with lib/heat and by the read_password.

Change-Id: If6a86dfaf0c21e2635c6de0a7b96a8ed7ec5b507