Updated openstack/openstack Project: openstack/horizon  93f81b4811952a896839a1148bec557af27d0254
Add Font Awesome icons

This patch adds an xstatic-packaged font awesome icon font to Horizon.
We need it, because the font icons included in Bootstrap 3 are non-free
and are likely to be excluded from some distributions. It also has a much
richer set of icons to choose from.

Change-Id: I725e0d77d230c16c01971f9af0dc02eb5662154d
Partial-implements: blueprint bp/font-icons