Updated openstack/openstack Project: openstack/nova  eccc42d4968bccafe85ee336cc2247d1eeaaf071
InstanceNUMAToplogy object create remove uuid param

We can get this from the object attribute so no need to pass it in
explicitly.

We do not need to bump the version of the object as there is no code
calling it as of yet, so we do not have to support the broken method
interface until the next major version bump.

Change-Id: I96e1d958e907272d94d7ca6481592b79ea6b6397
Blueprint: virt-driver-numa-placement