Updated openstack/openstack Project: openstack/tempest  dc2ef09a78bf1d1104c08c4dba8b7a1fb6435ff4
Add a check_telemetry method to javelin

Adds simple queries of the telemetry service that check that meter
data properly survives an upgrade.

JAVELIN_START is recorded to provide a way to confirm that the
oldest available data is from before the upgrade, thus confirming
that the data survived the upgrade.

Partially implements: blueprint grenade-resource-survivability

Change-Id: Ie759d71326916d543c7b3603d8812306d093715a