Updated openstack/openstack Project: openstack/trove  0fd90d54806ec742a7a79077dc9ce175759000a8
Fixed restore to work correctly with pexpect

Fixed an issue with restore where it was failing with the latest
version of pexpect in Ubuntu 14.04.

pexpect attempts to stop the temp mysqld process which fails due to
a difference in privilege level (since the process was run with sudo).
We now use mysqladmin shutdown to stop the temp mysqld process.

Closes-Bug: 1343344

Change-Id: Iba81648bb1654db97de876dd9a0dbfb04b8811d1