Updated openstack/openstack Project: openstack/taskflow  361ff11a1fc06a3423785049d2ae3acac5657807
Make the WBE worker banner information more meaningful

Add in more details that are displayed in the LOG when
a WBE worker is started up that is useful to show to be
able to help in debugging, or other informational and
operational purposes.

Example of the new output is the following:

TaskFlow v0.3.21.62 WBE worker.
Connection details:
  Driver = py-amqp v1.4.5
  Exchange = test
  Topic = test
  Transport = amqp
  Uri = amqp://guest@localhost:5672//
Powered by:
  Executor = concurrent.futures.thread.ThreadPoolExecutor
  Thread count = 3
Supported endpoints:
  - taskflow.tests.utils.NastyTask
  ...
  - taskflow.tests.utils.TaskMultiArgOneReturn
System details:
  Hostname = lappy.gateway.net
  Pid = 28364
  Platform = Linux-3.13.0-30-generic-x86_64-with-Ubuntu-14.04-trusty
  Python = 2.7.6 (default, Mar 22 2014, 22:59:56)
  Thread id = 139875992315712

Change-Id: I6d7dba3406007ddc80cce96cfdbbfd25935a12ae