Updated openstack/openstack Project: openstack/taskflow  722de97e3c7d2206867462c7363a0f072baff384
Allow worker count to be specified when no executor provided

When a multi-threaded engine is used it is nice to be able
to inform the engine how many workers should be created when
it creates its own executor. To allow this to be possible accept
a new keywork argument that can be used to set this value.

Change-Id: I0095d548249372440abbcf9b5c3b8fa841ca0ea9