Updated openstack/openstack Project: openstack/trove  343533504a728fe523140e7a8d771e1930ac8360
Cleaned up sample trove-guestagent.conf

Cleaned out the trove-guestagent.conf to only contain values that make
sense for the guest agent. There were also values that were different
between the sample conf, and the defaults in cfg.py, which were fixed
to be consistent, so as to have a single set of sane defaults.

Closes bug: 1336618

Change-Id: Ib3823856c288971665fa29489c4ffc5b899ccbaf