Updated openstack/openstack Project: openstack/nova  d8ea687c7cb1901ac9be36c850b4db301a6d90b4
tests: fix test_compute to have predictable service list

The _create_service_entries() method in test_compute.py accepted
a dict for the list of services. The tests would then pull out
the first item in the list of keys of the dict. When the python
hash seed is set to a random value the list of keys has no
predictable ordering, causing the test suite to break. Changing
the data structure to use a list of lists instead of a dict
ensures strict ordering.

Related-bug: #1350287
Change-Id: I01a0d4cfb086eb94b49d25b65daf08e55e2087b4