Updated openstack/openstack Project: openstack/horizon  b4d6a099df8ccebda2b54a77f8893941226c5144
Need ability to evacuate host in syspanel

Implement host evacuate in the hypervisors panel.
An extra tab was added to show in the first one hypervisors
and on the second one compute host. on each compute host
that is down an evacuate host button was added.
If the user press the button a modal windows is shown to
request the needed data to perform the evacuation.

blueprint evacuate-host

Co-Authored-By: Leandro Costantino <leandro.i.costantino@intel.com>
Co-Authored-By: David Lyle <david.lyle@hp.com>

Change-Id: I57a16f99fddd84c287429085c7e90beb59a17aa3