Updated openstack/openstack Project: openstack/cinder  0afe14b9b11d62e62cdcace4a09ff020a80e5229
Update the HP 3PAR default persona

The HP 3PAR drivers was setting the default persona to Generic.
The default should actually be Generic-ALUA for Linux based host.

DocImpact
This will require an update in the OpenStack Configuration Reference
Guide under the supported operations page for the HP 3PAR drivers.
The new default for the hp3par:persona should be '2 - Generic-ALUA'.

Change-Id: I65860a0b58f52ead0c86a0f0ab9a9ad755f76e72