Updated openstack/openstack Project: openstack/keystone  e2589177d9d8b2b31609e77619afd6ae462d13e3
Add test for getting a token with inherited role

If a user has an inherited role in a domain, he
is able to get a token on every project inside
that domain, even if he doesn't have a specific
grant on that project.
This patch adds a test for this behavior.

Change-Id: Ib65cb3443559d3bb9b7ffc2b96237072d2b5be56
Closes-Bug: 1365147