Updated openstack/openstack Project: openstack/heat  a6a3e9e9de8b247814f24255674a80d34fc504f9
Refactor nested stack create/update/preview

The bug #1357433 was caused by the fact that stack parsing for a stack
resource on create/update/preview is pretty similar but repeated in all
three places.  This refactors them to move common logic to a single
place, making sure all validation in consistent and all params are
passed in the same way.

Closes-Bug: #1357433
Change-Id: I20ee7f9a6dbf047798b0cc954aa861f774e683b9