Updated openstack/openstack Project: openstack-dev/pbr  4cecec5c788d78132f78c90d070a9ac5709d0884
Accept capitalized Sem-Ver headers

In Git, HTTP, and SMTP, header names generally have the first letter of
each word (including each part of a hyphenated word) capitalized.  In
all of the same cases, the actual headers are generally not considered
case-sensitive.  Accomodate this convention and the robustness principle
by performing lower-case comparisons of header names, and using
upper-case names in all of the documentation.

Change-Id: Iffa68107c8ff2aeaac92434da85155213e45a3cd