Updated openstack/openstack Project: openstack/horizon  fdf72d1a7c91d56dac74134a27532aeb14b74119
Add attributes to cell in a table depending on data

This feature helps users to selectively add attributes to cells
within a table based on the value in the cell. An example of why
this might be useful is cases where we want to display a short
description of what data in a given cell means.

This has been implemented by adding a new property to the Column
class called the 'cell_attributes_getter'. This attribute is a
callable which returns a dictionary with the html attributes and
their values. The attribute defined in this dictionary can include
any HTML global attributes

Change-Id: I9e9aa1bc32e7afd1a62a69dc13ef0ed0d0e748c4
Implements: blueprint add-attributes-for-table-cells