Updated openstack/openstack Project: openstack/horizon  2f9c1b37e664d42b9020b7c6d0d0cb802959b99d
Fixed incorrect position of modal container.

After updating to bootrap 3, the position of the modal
container is incorrect when you scroll down on a page
then select an action button towards the top of the page that brings up
a modal dialog.  The dialog is displayed off the bottom of the screen.

I couldnt find a graceful way to fix the css issue, so I had to over-
ride a bootstrap css rule using !important.

Change-Id: I5539692a6657555e08e893339098159306836021
Closes-bug: #1353897