Updated openstack/openstack Project: openstack/oslo-incubator  6b048e79e88122f09697d2b194cf266ef267bf9c
Let oslotest manage the six.move setting for mox

This instruction to set up a move in six for the mox module conflicts
with the one in oslotest. We are trying to use mox3 everywhere, so
remove this instruction and let oslotest handle it.

This changeset modifies the tests for the old qpid RPC driver, which is
deprecated. However, we need to make the file importable for testr to
successfully run the tests that aren't deprecated, so we need that
change as well.

Change-Id: I59d5799283233f8411044ddb15c8abfc8850014c