Updated openstack/openstack Project: openstack/oslo-incubator  eb6e8824e5235316707b7ba81aea8037385b6114
Test formatting errors with log level being emitted

The test to ensure that formatting errors are handled properly was being
run using info level logging, but the default log configuration for
tests does not emit messages at that level, so switch to error.

Back-ported from commit with the same ID in oslo.log.

Change-Id: Ie11a51deea65627b45a11d7dfca36d16c1b5949e